+++ examples/combo_box/disabled/index.jsx
@@ -1,0 +1,45 @@
+import React, { useState } from 'react';
+import { ComboBox, useTokenSearch, tokenHighlight } from '../../../src';
+
+const options = [
+  'Apple',
+  'Orange',
+  'Blood orange',
+  'Lemon',
+  'Raspberry',
+  'Strawberry',
+];
+
+export function Example() {
+  const [value, setValue] = useState('Lemon');
+  const [search, setSearch] = useState(null);
+  const filteredOptions = useTokenSearch(search, { options });
+
+  return (
+    <>
+      <label
+        id="select-label"
+        htmlFor="select"
+      >
+        Select
+      </label>
+      <ComboBox
+        id="select"
+        aria-labelledby="select-label"
+        value={value}
+        onValue={setValue}
+        onSearch={setSearch}
+        options={filteredOptions}
+        renderValue={tokenHighlight()}
+        disabled
+      />
+
+      <label htmlFor="output">
+        Current value
+      </label>
+      <output htmlFor="select" id="output">
+        {JSON.stringify(value, undefined, ' ')}
+      </output>
+    </>
+  );
+}